schmivitz: update witness counter to track witness
In order to commit to the extended witness, we need to track exactly
which intermediate values are part of it, not just the number of them.

+++ schmivitz/src/prove.rs
@@ -35,7 +35,7 @@ impl Proof {
         println!(
             "The total number of required VOLEs is {}",
             extended_witness_counter.count()
-                + (parameters::VOLE_SIZE_PARAM * parameters::REPETITION_PARAM) as u64
+                + (parameters::VOLE_SIZE_PARAM * parameters::REPETITION_PARAM)
         );
 
         todo!()
